+++ cmd/flowg-client/cmd_acl.go
@@ -1,0 +1,17 @@
+package main
+
+import "github.com/spf13/cobra"
+
+func NewAclCommand() *cobra.Command {
+	cmd := &cobra.Command{
+		Use:   "acl",
+		Short: "Access Control commands",
+	}
+
+	cmd.AddCommand(
+		NewAclUserCommand(),
+		NewAclRoleCommand(),
+	)
+
+	return cmd
+}
