+++ cmd/flowg-client/cmd_acl_role.go
@@ -1,0 +1,20 @@
+package main
+
+import "github.com/spf13/cobra"
+
+func NewAclRoleCommand() *cobra.Command {
+	cmd := &cobra.Command{
+		Use:   "role",
+		Short: "Manage roles",
+	}
+
+	cmd.AddCommand(
+		NewAclRoleListCommand(),
+		NewAclRoleAddCommand(),
+		NewAclRoleDeleteCommand(),
+		NewAclRoleGrantCommand(),
+		NewAclRoleRevokeCommand(),
+	)
+
+	return cmd
+}
