+++ cmd/flowg-client/cmd_acl_user.go
@@ -1,0 +1,20 @@
+package main
+
+import "github.com/spf13/cobra"
+
+func NewAclUserCommand() *cobra.Command {
+	cmd := &cobra.Command{
+		Use:   "user",
+		Short: "Manage users",
+	}
+
+	cmd.AddCommand(
+		NewAclUserListCommand(),
+		NewAclUserAddCommand(),
+		NewAclUserDeleteCommand(),
+		NewAclUserGrantCommand(),
+		NewAclUserRevokeCommand(),
+	)
+
+	return cmd
+}
