+++ cmd/flowg-client/cmd_acl_user_grant.go
@@ -12,32 +12,24 @@ import (
 
 	"link-society.com/flowg/api"
 
-	"link-society.com/flowg/internal/models"
 	"link-society.com/flowg/internal/utils/client"
 )
 
-func NewAdminRoleRevokeCommand() *cobra.Command {
+func NewAclUserGrantCommand() *cobra.Command {
 	type options struct {
-		name       string
-		permission string
+		username string
+		rolename string
 	}
 
 	opts := &options{}
 
 	cmd := &cobra.Command{
-		Use:   "revoke",
-		Short: "Revoke a permission from a role",
+		Use:   "grant",
+		Short: "Grant a role to a user",
 		Run: func(cmd *cobra.Command, args []string) {
-			oldScope, err := models.ParseScope(opts.permission)
-			if err != nil {
-				fmt.Fprintf(os.Stderr, "ERROR: Invalid permission: %v\n", err)
-				exitCode = 1
-				return
-			}
-
 			client := cmd.Context().Value(ApiClient).(*client.Client)
 
-			url := fmt.Sprintf("/api/v1/roles/%s", opts.name)
+			url := fmt.Sprintf("/api/v1/users/%s", opts.username)
 			req, err := http.NewRequest(http.MethodGet, url, nil)
 			if err != nil {
 				fmt.Fprintf(os.Stderr, "ERROR: Could not prepare request: %v\n", err)
@@ -59,7 +51,7 @@ func NewAdminRoleRevokeCommand() *cobra.Command {
 				return
 			}
 
-			var data api.GetRoleResponse
+			var data api.GetUserResponse
 			if err := json.NewDecoder(resp.Body).Decode(&data); err != nil {
 				fmt.Fprintf(os.Stderr, "ERROR: Could not decode response: %v\n", err)
 				exitCode = 1
@@ -67,27 +59,19 @@ func NewAdminRoleRevokeCommand() *cobra.Command {
 			}
 
 			found := false
-			for i, scope := range data.Role.Scopes {
-				if scope == oldScope {
-					data.Role.Scopes = append(
-						data.Role.Scopes[:i],
-						data.Role.Scopes[i+1:]...,
-					)
+			for _, role := range data.User.Roles {
+				if role == opts.rolename {
 					found = true
 					break
 				}
 			}
-
-			if found {
-				scopes := make([]string, len(data.Role.Scopes))
-				for i, scope := range data.Role.Scopes {
-					scopes[i] = string(scope)
-				}
+			if !found {
+				data.User.Roles = append(data.User.Roles, opts.rolename)
 
 				body := struct {
-					Scopes []string `json:"scopes"`
+					Roles []string `json:"roles"`
 				}{
-					Scopes: scopes,
+					Roles: data.User.Roles,
 				}
 
 				payload, err := json.Marshal(body)
@@ -97,7 +81,7 @@ func NewAdminRoleRevokeCommand() *cobra.Command {
 					return
 				}
 
-				req, err := http.NewRequest(http.MethodPut, url, bytes.NewBuffer(payload))
+				req, err := http.NewRequest(http.MethodPatch, url, bytes.NewBuffer(payload))
 				if err != nil {
 					fmt.Fprintf(os.Stderr, "ERROR: Could not prepare request: %v\n", err)
 					exitCode = 1
@@ -122,20 +106,20 @@ func NewAdminRoleRevokeCommand() *cobra.Command {
 	}
 
 	cmd.Flags().StringVar(
-		&opts.name,
-		"name",
+		&opts.username,
+		"username",
 		"",
-		"Name of the role to revoke the permission from",
+		"Name of the user to grant the role to",
 	)
-	cmd.MarkFlagRequired("name")
+	cmd.MarkFlagRequired("username")
 
 	cmd.Flags().StringVar(
-		&opts.permission,
-		"permission",
+		&opts.rolename,
+		"rolename",
 		"",
-		"Permission to revoke from the role",
+		"Name of the role to grant to the user",
 	)
-	cmd.MarkFlagRequired("permission")
+	cmd.MarkFlagRequired("rolename")
 
 	return cmd
 }
