+++ cmd/flowg-client/cmd_admin.go
@@ -11,8 +11,6 @@ func NewAdminCommand() *cobra.Command {
 	cmd.AddCommand(
 		NewAdminBackupCommand(),
 		NewAdminRestoreCommand(),
-		NewAdminUserCommand(),
-		NewAdminRoleCommand(),
 	)
 
 	return cmd
