+++ cmd/flowg-client/main.go
@@ -58,6 +58,7 @@ func main() {
 		NewStreamCommand(),
 		NewPipelineCommand(),
 		NewTransformerCommand(),
+		NewAclCommand(),
 		NewAdminCommand(),
 	)
